App files moved to src folder
